imp - Nowrap is disabled on alert boxes
We need to wrap text in alert boxes.

---

To wrap text in alert boxes, we need to remove the nowrap class in both Alert and ContactBox in order to be able to wrap text in small screens, especially when it comes to mobile devices.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1